requirements.txt
Django==5.2.4
Faker==37.4.2
django-ninja==1.4.3
